appmenus: support per-vm appmenus templates
Add support for multiple appmenus templates dir. Specifically,
TemplateBasedVM can have own appmenus templates (for example extracted
from ~/.local/share/applications), which extend/override those from
the template. Technically, each VM have now a list of appmenus template
directories (not a single dir), which are searched in relevance order
(up in 'template' hierarhy).

This is especially useful if one install and application in
TemplateBasedVM for example in user home or /usr/local. This makes it
easier to add such application to the menu.

For this change to work properly, there is a need for change in
qubes.GetAppmenus service on the VM side, to not report applications
installed on /, if changes in / are not persistent there (i.e.
TemplateBasedVM). Otherwise applications installed in templates will be
retrieved multiple times, wasting time and disk space. But updating VM
later should clean this up.

Fixes QubesOS/qubes-issues#4152