Switch appmenu delimiter from hyphen to full stop
Full stops (‘.’) are not permitted in qube names, so this avoids
ambiguity.  We also use org.qubes-os prefixes for desktop names, as
recommended by the FreeDesktop.org specification.  There is one
exception: the settings desktop file now has an ID beginning with
‘qubes-vm-settings’, which prevents it from clashing with any
application in any qube.

The one exception is menu directory names.  ‘xdg-desktop-menu’ merges
any menus that are identical up to the first period.  Therefore, we must
ensure that the prefix preceding the qube name has no periods in it.
This doesn’t create an unsolvable conflict, however, because QubesOS VM
menu directories have no data that is under an attacker’s control.

When ‘qubes-desktop-linux-common’ is upgraded, all of the menus need to
be removed and replaced.  However, ‘xdg-desktop-menu’ gets confused if
entries in two or more menus are removed simultaneously.  Prevent this
by submitting two lists of files for removal: one containing files that
use the old naming scheme, and one containing files using the new one.