Only allow DispVM templates to have appmenus-dispvm feature
If the appmenus-dispvm feature is set on a VM that is not a DispVM
template, old versions of the app menu code get confused and try to use
the VM as a DispVM template.  Having the appmenus-dispvm feature set on
a VM that is not a DispVM template makes no sense, so reject it
outright.